+++ Sources/KeyboardKit/UIKit/Components/UIKeyboardButtonRowComponent.swift
@@ -1,5 +1,5 @@
 //
-//  KeyboardButtonRowComponent.swift
+//  UIKeyboardButtonRowComponent.swift
 //  KeyboardKit
 //
 //  Created by Daniel Saidi on 2019-04-28.
@@ -12,4 +12,4 @@ import UIKit
  This protocol represents a view component that can be added
  to a keyboard button row.
  */
-public protocol KeyboardButtonRowComponent: HorizontalKeyboardComponent {}
+public protocol UIKeyboardButtonRowComponent: UIHorizontalKeyboardComponent {}