+++ Sources/KeyboardKit/UIKit/KeyboardSpacerView.swift
@@ -14,7 +14,7 @@ import UIKit
  You could for instance map `.none` actions to this view, to
  let you declare "spaces" in your action collections.
  */
-open class KeyboardSpacerView: UIView, KeyboardButtonRowComponent {
+open class KeyboardSpacerView: UIView, UIKeyboardButtonRowComponent {
     
     public convenience init(width: CGFloat) {
         self.init(frame: .zero)